requirements.txt
Flask==2.0.0
SQLAlchemy>=1.4,<2.0
cryptography
scikit-learn
litellm
openai
python-telegram-bot[job-queue]==22.0
python-dotenv


Flask>=2.0.0,<3.0.0
requests==2.31.0








